Refactors CLI to use argparse and adds image download functionality
Migrates the CLI from Click to `argparse` for improved flexibility and maintainability.

Adds the ability to download card images with specified sizes (small, normal, large) via the `--get-images` argument.

Updates the image download logic to support different image sizes.
